Consolidate GUI and plan runner
