--- json_plan_example.json
@@ -1,22 +1,0 @@
-{
-  "modules": [
-    {
-      "name": "Demo",
-      "tests": [
-        {
-          "name": "Simple",
-          "steps": [
-            {
-              "description": "Write and verify INT",
-              "db_number": 1,
-              "start": 0,
-              "data_type": "INT",
-              "write": 123,
-              "expected": 123
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
