+++ plc_tester_gui.py
@@ -5,13 +5,12 @@ A simple GUI to create and run PLC test plans using Snap7.
 import json
 import time
 from dataclasses import dataclass, field, asdict
-from typing import Any, List
+from typing import Any, Dict, List, Union
 
 import tkinter as tk
 from tkinter import filedialog, messagebox, simpledialog, ttk
 
 import snap7
-from plan_runner import run_json_plan as execute_plan
 from snap7.util import (
     get_bool,
     get_dint,
@@ -81,6 +80,95 @@ TYPE_FUNCS = {
     "DWORD": (4, set_dword, get_dword),
     "REAL": (4, set_real, get_real),
 }
+
+
+def run_plan(plan: TestPlan, ip: str = "127.0.0.1", rack: int = 0, slot: int = 1) -> None:
+    """Run ``plan`` against the specified PLC."""
+    conn = PLCConnection()
+    conn.connect(ip, rack, slot)
+    try:
+        for module in plan.modules:
+            for test in module.tests:
+                _run_test(conn, test)
+    finally:
+        conn.disconnect()
+
+
+def run_json_plan(
+    data: Union[str, Dict[str, Any]],
+    ip: str = "127.0.0.1",
+    rack: int = 0,
+    slot: int = 1,
+) -> None:
+    """Run a test plan described as JSON."""
+
+    if isinstance(data, str):
+        with open(data, "r", encoding="utf-8") as fh:
+            data = json.load(fh)
+    plan = TestPlan.from_dict(data)
+    run_plan(plan, ip=ip, rack=rack, slot=slot)
+
+
+def _run_test(conn: "PLCConnection", test: TestCase) -> None:
+    """Execute a single :class:`TestCase` using ``conn``."""
+
+    for step in test.steps:
+        starts = step.start if isinstance(step.start, list) else [step.start]
+        types = step.data_type if isinstance(step.data_type, list) else [step.data_type]
+        if len(types) == 1 and len(starts) > 1:
+            types = types * len(starts)
+        writes = step.write if isinstance(step.write, list) else ([step.write] if step.write is not None else [])
+        expecteds = (
+            step.expected if isinstance(step.expected, list) else ([step.expected] if step.expected is not None else [])
+        )
+
+        if step.delay_ms:
+            time.sleep(step.delay_ms / 1000.0)
+
+        for idx, start in enumerate(starts):
+            dtype = types[idx]
+            w = writes[idx] if idx < len(writes) else None
+            e = expecteds[idx] if idx < len(expecteds) else None
+            if dtype == "BOOL":
+                byte_str, bit_str = str(start).split(".")
+                byte_idx, bit_idx = int(byte_str), int(bit_str)
+                if w is not None:
+                    cur = bytearray(conn.read(step.db_number, byte_idx, 1))
+                    set_bool(cur, 0, bit_idx, bool(w))
+                    conn.write(step.db_number, byte_idx, bytes(cur))
+                if e is not None:
+                    data = conn.read(step.db_number, byte_idx, 1)
+                    val = get_bool(data, 0, bit_idx)
+                    if val != bool(e):
+                        raise AssertionError(
+                            f"{step.description} at {start}: expected {e} got {val}"
+                        )
+            elif dtype == "BYTE":
+                addr = int(start)
+                if w is not None:
+                    conn.write(step.db_number, addr, bytes([int(w)]))
+                if e is not None:
+                    data = conn.read(step.db_number, addr, 1)
+                    val = data[0]
+                    if val != int(e):
+                        raise AssertionError(
+                            f"{step.description} at {start}: expected {e} got {val}"
+                        )
+            else:
+                size, set_func, get_func = TYPE_FUNCS[dtype]
+                addr = int(start)
+                if w is not None:
+                    buf = bytearray(size)
+                    set_func(buf, 0, w)
+                    conn.write(step.db_number, addr, bytes(buf))
+                if e is not None:
+                    data = conn.read(step.db_number, addr, size)
+                    val = get_func(data, 0)
+                    ok = abs(val - float(e)) < 1e-6 if dtype == "REAL" else val == e
+                    if not ok:
+                        raise AssertionError(
+                            f"{step.description} at {start}: expected {e} got {val}"
+                        )
 
 
 class StepEditor(tk.Toplevel):
@@ -272,7 +360,7 @@ class PlanJsonEditor(tk.Toplevel):
         except Exception as exc:  # pragma: no cover - user input
             messagebox.showerror("JSON error", str(exc))
             return
-        execute_plan(
+        run_json_plan(
             data,
             ip=self.gui.ip_var.get(),
             rack=self.gui.rack_var.get(),
